requirements.txt
fastapi~=0.116.0
PyYAML~=6.0.2
uvicorn~=0.35.0
psycopg2-binary==2.9.11
bcrypt==5.0.0
pydantic[email]
PyJWT==2.10.1
pydantic-settings==2.12.0
python-multipart==0.0.20